feat(chatbot): sseparser
